+++ flatlaf-demo/src/main/java/com/formdev/flatlaf/demo/DemoFrame.java
@@ -250,7 +250,7 @@ class DemoFrame
 				titleLabel,
 				"Demonstrates FlatLaf Swing look and feel",
 				" ",
-				"Copyright 2019-" + Year.now() + " FormDev Software GmbH",
+				"Copyright 2019-" + Year.now() + " FormDev Software",
 				linkLabel,
 			},
 			"About", JOptionPane.PLAIN_MESSAGE );